cobo-tssnode-callback/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
blinker==1.9.0
cffi==2.0.0
click==8.3.0
cobo-waas2==1.15.0
cryptography==44.0.1
cytoolz==1.0.1
dotenv==0.9.9
ecdsa==0.19.0
eth-hash==0.7.1
eth-typing==5.2.1
eth-utils==5.3.1
fastapi==0.103.2
Flask==3.1.1
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pika==1.3.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.2
pydantic_core==2.41.4
PyJWT==2.10.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
toolz==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.0.7
uvicorn==0.22.0
Werkzeug==3.1.3
